Add sort_order field to goal, and refactor Student component for upcoming changes
